packages/nonebot-plugin-nailongremove-base/pyproject.toml
[project]
name = "nonebot-plugin-nailongremove-base"
dynamic = ["version"]
description = "识别目标并撤回图片插件"
authors = [
    { name = "445", email = "2877834692@qq.com" },
    { name = "student_2333", email = "lgc2333@126.com" },
]
requires-python = "<4.0,>=3.9"
dependencies = [
    "nonebot2>=2.2.0",
    "nonebot-plugin-alconna>=0.53.1",
    "nonebot-plugin-uninfo>=0.5.0",
    "opencv-python>=4.5",
    "numpy>=1.19",
    "pillow>=9",
    "cookit[pydantic]>=0.8.1",
    "httpx>=0.27.2",
    "githubkit>=0.11.14",
    "yarl>=1.17.1",
    "tqdm>=4.66.6",
    "huggingface-hub>=0.26",
    "ultralytics>=8.3",
    "gradio_client>=1.4"

]
license = { text = "MIT" }
readme = "README.md"
keywords = ["nonebot", "plugin", "image"]

[project.urls]
homepage = "https://github.com/Refound-445/nonebot-plugin-nailongremove"
repository = "https://github.com/Refound-445/nonebot-plugin-nailongremove"

[project.optional-dependencies]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.build]
includes = []

[tool.pdm.version]
source = "file"
path = "nonebot_plugin_nailongremove/__init__.py"
